change dummy to noauth mode.
